backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.32.1
google-generativeai==0.8.6
python-dotenv==1.0.0
python-multipart==0.0.9
pillow==10.2.0
